CDKPackage/requirements.txt
aws-cdk-lib>=2,<3
constructs>=10,<11
